video-scope-analyzer/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
ffmpeg-python>=0.2.0
python-docx>=0.8.11
reportlab>=3.6.0
google-api-python-client>=2.80.0
google-auth>=2.23.4
websockets>=12.0
aiofiles>=23.2.1
redis>=5.0.0
rq>=1.15.1